adds profile user ranking
